// src/app/core/enum/message.enum.ts
export enum MessageID {    
    ComponentsDetails = 1,
    ComponentsReleaseDetails = 2,
    SubscribeReleaseMail = 3,
    UnsubscribeReleaseMail = 4,
    
    Component1Details = 5,
    Component2Details = 6,
    Component3ComponentDetails = 7,
    
    Component1ComponentReleaseDetails = 8,
    Component2ComponentReleaseDetails = 9,
    Component3ComponentReleaseDetails = 10,

    Component1Subscribe = 11,
    Component2Subscribe = 12,
    Component3Subscribe = 13,

    Component1Unsubscribe = 14,
    Component2Unsubscribe = 15,
    Component3Unsubscribe = 16,
}